🐛 Refactor to remove shared code between admin and non-admin for Genomic Variants * Refactor all genomic variant import notification code to the admin-only NotificationsContext * Split removeImportNotification into separate functions for removing the import notification and showing the successful import notification
