+++ ui/src/components/admin/Notifications/NotificationsController.js
@@ -1,9 +1,12 @@
-import React, { useState } from 'react';
+import React, { useContext, useState } from 'react';
+
+import NOTIFICATION_TYPES from './NotificationTypes';
 
 export const NotificationsContext = React.createContext();
 
 const NotificationsProvider = ({ children }) => {
   const [notifications, setNotifications] = useState([]);
+  const [importNotifications, setImportNotifications] = useState([]);
 
   const appendNotification = notification => {
     const id = Date.now();
@@ -50,11 +53,77 @@ const NotificationsProvider = ({ children }) => {
         setNotifications,
         appendNotification,
         clearNotification,
+        importNotifications,
+        setImportNotifications,
       }}
     >
       {children}
     </NotificationsContext.Provider>
   );
+};
+
+export const useGenomicVariantImportNotifications = () => {
+  const {
+    appendNotification,
+    clearNotification,
+    importNotifications,
+    setImportNotifications,
+  } = useContext(NotificationsContext);
+
+  const getImportNotifications = () => {
+    if (!importNotifications) return [];
+    return importNotifications;
+  };
+
+  const addImportNotification = async modelName => {
+    const existingNotification = getImportNotifications().find(x => x.modelName === modelName);
+
+    if (!existingNotification) {
+      const notification = await appendNotification({
+        type: NOTIFICATION_TYPES.LOADING,
+        message: `Importing: Research Variants for ${modelName} are currently importing.`,
+        details:
+          'You can continue to use the CMS, and will be notified when the import is complete.',
+        timeout: false,
+      });
+
+      setImportNotifications(notifications => [
+        ...notifications,
+        {
+          modelName: modelName,
+          notificationId: notification.id,
+        },
+      ]);
+    }
+  };
+
+  const removeImportNotification = modelName => {
+    const existingNotification = getImportNotifications().find(x => x.modelName === modelName);
+
+    if (existingNotification) {
+      clearNotification(existingNotification.notificationId);
+      setImportNotifications(notifications =>
+        notifications.filter(notification => notification.modelName !== modelName),
+      );
+    }
+  };
+
+  const showSuccessfulImportNotification = modelName => {
+    appendNotification({
+      type: NOTIFICATION_TYPES.SUCCESS,
+      message: `Import Successful: Research Somatic Variants for ${modelName} have successfully imported, however are not yet published.`,
+      link: `/admin/model/${modelName}`,
+      linkText: 'View on model page to publish these variants.',
+    });
+  };
+
+  return {
+    importNotifications: getImportNotifications(),
+    setImportNotifications,
+    addImportNotification,
+    removeImportNotification,
+    showSuccessfulImportNotification,
+  };
 };
 
 export default NotificationsProvider;
