+++ ui/src/components/admin/Notifications/index.js
@@ -1,3 +1,7 @@
 export { default as NotificationToaster } from './NotificationToaster';
-export { default as NotificationsProvider, NotificationsContext } from './NotificationsController';
+export {
+  default as NotificationsProvider,
+  NotificationsContext,
+  useGenomicVariantImportNotifications,
+} from './NotificationsController';
 export { default as NOTIFICATION_TYPES } from './NotificationTypes';
